Module twilight_gateway::cluster

The cluster is a manager for running and maintaining multiple shards, bringing their event streams into one unified stream.

Examples

Start a cluster of 10 shards and print when a shard is connected, disconnected, and when new message commands come in:

use futures::StreamExt;
use std::{env, sync::Arc};
use twilight_gateway::{Cluster, Event, Intents};

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let token = env::var("DISCORD_TOKEN")?;
    let intents =
        Intents::GUILD_BANS | Intents::GUILD_EMOJIS_AND_STICKERS | Intents::GUILD_MESSAGES;
    let (cluster, mut events) = Cluster::new(token, intents).await?;
    let cluster = Arc::new(cluster);
    cluster.up().await;

    while let Some((shard_id, event)) = events.next().await {
        tokio::spawn(handle_event(Arc::clone(&cluster), shard_id, event));
    }

    println!("Cluster is now shutdown");

    Ok(())
}

async fn handle_event(cluster: Arc<Cluster>, shard_id: u64, event: Event) {
    match event {
        Event::ShardConnected { .. } => {
            println!("Shard {shard_id} is now connected");
        }
        Event::ShardDisconnected { .. } => {
            println!("Shard {shard_id} is now disconnected");
        }
        Event::MessageCreate(msg) if msg.content == "!latency" => {
            if let Some(shard) = cluster.shard(shard_id) {
                if let Ok(info) = shard.info() {
                    println!("Shard {shard_id}'s latency is {:?}", info.latency());
                }
            }
        }
        Event::MessageCreate(msg) if msg.content == "!shutdown" => {
            println!("Got a shutdown request from shard {shard_id}");

            cluster.down();
        }
        _ => {}
    }
}

Start bucket 1 of a very large bot with 320 shards and a maximum concurrency of 16:

use futures::StreamExt;
use std::env;
use twilight_gateway::{cluster::ShardScheme, Cluster, Event, Intents};

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let token = env::var("DISCORD_TOKEN")?;
    let intents = Intents::GUILD_MESSAGES;
    let scheme = ShardScheme::try_from((1, 16, 320))?;
    let (cluster, mut events) = Cluster::builder(token, intents)
        .shard_scheme(scheme)
        .build()
        .await?;

    cluster.up().await;

    while let Some((shard_id, event)) = events.next().await {
        println!("got event type {:?}", event.kind());
    }

    println!("Cluster is now shutdown");

    Ok(())
}

Re-exports

pub use self::scheme::ShardScheme;

pub use self::scheme::ShardSchemeRangeError;

pub use self::scheme::ShardSchemeRangeErrorType;

Modules

scheme

Schemes for instantiating a cluster of shards.

Structs

Cluster

A manager for multiple shards.

ClusterBuilder

Builder to configure and construct a Cluster.

ClusterCommandError

Sending a command to a shard failed.

ClusterStartError

Starting a cluster failed.

Config

Built configuration for a Cluster.

Events

Stream of events from a Cluster.

Shards

Iterator over a Cluster’s managed shards.

Enums

ClusterCommandErrorType

Type of ClusterCommandError that occurred.

ClusterStartErrorType

Type of ClusterStartErrorType that occurred.