Enum twilight_gateway::cluster::scheme::ShardScheme

#[non_exhaustive]pub enum ShardScheme {
    Auto,
    Bucket {
        bucket_id: u64,
        concurrency: u64,
        total: u64,
    },
    Range {
        from: u64,
        to: u64,
        total: u64,
    },
}

The method of sharding to use.

By default this is Auto.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Auto

Specifies to retrieve the amount of shards recommended by Discord and then start all of them.

For example, if Discord recommends 10 shards, then all 10 shards will be started.

Bucket

Manage a single bucket’s worth of shards within the cluster.

This is primarily useful for bots in the Sharding for Very Large Bots program.

Fields of Bucket

bucket_id: u64

ID of the first shard to start.

This must be less than the maximum concurrency.

For example, if you have a maximum concurrency of 16 and the bucket ID is 0, then shards 0, 16, 32, etc. will be managed by the cluster.

concurrency: u64

Number of shards allowed to be started simultaneously within a bucket, also known as the maximum concurrency.

This is provided via SessionStartLimit::max_concurrency.

total: u64

Total number of shards used across all clusters.

Range

Specifies to start a range of shards.

Examples

For example, if your bot uses 50 shards, then you might specify to start shards 0 through 24:

use twilight_gateway::cluster::ShardScheme;
use std::convert::TryFrom;

let range = ShardScheme::try_from((0..24, 50))?;

Fields of Range

from: u64

First shard ID to spawn.

to: u64

Last shard ID to spawn.

This doesn’t necessarily have to be up to the total.

total: u64

Total number of shards used by the bot.

Implementations

impl ShardScheme

pub fn iter(&self) -> Option<ShardSchemeIter>

Consume the shard scheme, returning an iterator of the shards that it denotes.

Returns None if the scheme is dynamic, i.e. the scheme is the Auto variant.

pub fn from(&self) -> Option<u64>

First shard ID that will be started, if known.

In the case of the Auto variant the total is unknown.

pub fn total(&self) -> Option<u64>

Total number of shards used by the bot across all clusters, if known.

In the case of the Auto variant the total is unknown.

pub fn to(&self) -> Option<u64>

Maximum shard ID across all clusters, if known.

In the case of the Auto variant the total is unknown.