Struct twilight_gateway::cluster::ClusterBuilder

pub struct ClusterBuilder { /* private fields */ }

Builder to configure and construct a Cluster.

Examples

Create a cluster with only the GUILD_MESSAGES intents with a large_threshold of 100.

use std::env;
use twilight_gateway::{Cluster, Intents};

let token = env::var("DISCORD_TOKEN")?;

let cluster = Cluster::builder(token, Intents::GUILD_MESSAGES)
    .large_threshold(100)
    .build()
    .await?;

Implementations

impl ClusterBuilder

pub fn new(token: String, intents: Intents) -> Self

Create a new builder to construct and configure a cluster.

pub async fn build(self) -> Result<(Cluster, Events), ClusterStartError>

Consume the builder and create the cluster.

Errors

Returns a ClusterStartErrorType::RetrievingGatewayInfo error type if there was an HTTP error Retrieving the gateway information.

pub fn event_types(self, event_types: EventTypeFlags) -> Self

Set the event types to process.

This is an optimization technique; all events not included in the provided event type flags will not be deserialized by the gateway and will be discarded. All events will still be sent if EventTypeFlags::SHARD_PAYLOAD is enabled.

pub fn gateway_url(self, gateway_url: String) -> Self

Set the URL that will be used to connect to the gateway.

Default is to fetch it from the HTTP API.

pub fn http_client(self, http_client: Arc<Client>) -> Self

Set the twilight_http Client used by the cluster and the shards it manages.

This is needed so that the cluster and shards can retrieve gateway information.

Defaults to a new, default HTTP client is used.

pub fn identify_properties(
self,
identify_properties: IdentifyProperties
) -> Self

Set the properties for shards to identify with.

This may be used if you want to set a different operating system, for example.

Examples

Set the identify properties for a cluster:

use std::env::{self, consts::OS};
use twilight_gateway::{Intents, Cluster};
use twilight_model::gateway::payload::outgoing::identify::IdentifyProperties;

let token = env::var("DISCORD_TOKEN")?;
let properties = IdentifyProperties::new("twilight.rs", "twilight.rs", OS);

let builder = Cluster::builder(token, Intents::empty())
    .identify_properties(properties);

pub fn large_threshold(self, large_threshold: u64) -> Self

Set the “large threshold” of shards.

Refer to the shard’s ShardBuilder::large_threshold for more information.

Panics

Panics if the provided value is below 50 or above 250.

pub fn presence(self, presence: UpdatePresencePayload) -> Self

Set the presence to use when identifying with the gateway.

Refer to the shard’s ShardBuilder::presence for more information.

pub fn ratelimit_payloads(self, ratelimit_payloads: bool) -> Self

Set whether or not outgoing payloads will be ratelimited.

Useful when running behind a proxy gateway. Running without a functional ratelimiter will get you ratelimited.

Defaults to being enabled.

pub fn shard_presence<F>(self, shard_presence: F) -> Self where
F: Fn(u64) -> Option<UpdatePresencePayload> + Send + Sync + 'static,

Set specific shard presences to use when identifying with the gateway.

Accepts a closure. The closure accepts a u64 and returns an Option<UpdatePresencePayload>. This presence will override any set by presence, even if the provided closure returns None.

pub fn shard_scheme(self, scheme: ShardScheme) -> Self

Set the scheme to use for shard managing.

ShardScheme::Range means that it will manage a range of shards, but not necessarily all of the shards that your bot uses.

The cluster will automatically manage all of the shards that Discord recommends you use by default. For most setups this is an acceptable default.

Examples

Configure a cluster to manage shards 0-9 out of 20 shards total:

use twilight_gateway::{cluster::{Cluster, ShardScheme}, Intents};
use std::env;

let token = env::var("DISCORD_TOKEN")?;
let scheme = ShardScheme::try_from((0..=9, 20))?;

let cluster = Cluster::builder(token, Intents::GUILD_MESSAGES)
    .shard_scheme(scheme)
    .build()
    .await?;

pub fn queue(self, queue: Arc<dyn Queue>) -> Self

Set the queue to use for queueing shard connections.

This is useful when you have a very large bot or when you have a more advanced setup with multiple processes connecting at the same time.

Refer to the queue module for more information.

pub fn resume_sessions(
self,
resume_sessions: HashMap<u64, ResumeSession>
) -> Self

Set the session information to resume shards with.

This requires having recovered the resume data when shutting down the cluster via Cluster::down_resumable.

Note that this does not guarantee all or any of the shards will be able to resume. If their sessions are invalid they will have to re-identify to initialize a new session.