Module twilight_gateway::shard::raw_message

Send raw websocket messages over the websocket.

This is mostly equivalent to the underlying websocket library’s message, but this intermediary exists to prevent exposing it in the public API. Messages constructed are equivalent to what the underlying library will receive. The input will not be checked and will be passed directly to the underlying websocket library.

Structs

CloseFrame
Information about a close message, if any.

Enums

Message
Message to send over the connection to the remote.