Module twilight_gateway::shard::stage

Utilities for knowing and parsing the current connection stage of a Shard.

Included is the Stage, which is an enum representing the connection stage with variants such as Connecting or Disconnected.

The Stage also has some parsing capability, so an error type for conversion reasons is included.

Enums

Stage
The current connection stage of a Shard.

StageConversionError
Reason for a failure while parsing a value into a Stage.