Struct twilight_gateway::stream::ShardRef

pub struct ShardRef<'a> { /* private fields */ }

Guard dereferencing to the shard that produced an event or message.

Note that manually causing the destructor to not be called will cause the shard to not be re-inserted into the stream.