Struct twilight_gateway::EventTypeFlags

pub struct EventTypeFlags(/* private fields */);

Important optimization for narrowing requested event types.

Specifying event types is an important optimization technique on top of Intents, which can dramatically decrease processor usage in many circumstances. While specifying intents are required by Discord and allow filtering groups of Events, event type flags are a Twilight-specific technique to filter out individual events from being deserialized at all, effectively discarding them.

For example, Intents::GUILDS includes a wide range of events from GuildCreate to RoleUpdate to ChannelPinsUpdate. If the only events used in this group of events is, say, ChannelCreate and RoleCreate, then the CHANNEL_CREATE and ROLE_CREATE event type flags can be specified in combination with that intent to only deserialize those events.

Selected event types only affect the events returned by Shards. Events necessary for maintaining the connection to Discord, such as GATEWAY_HEARTBEAT and GATEWAY_HELLO, can safely be excluded and won’t cause the operation of shards to fail, because shards will always parse portions of necessary events.

Implementations

impl EventTypeFlags

pub const AUTO_MODERATION_ACTION_EXECUTION: Self = _

Message has been blocked by AutoMod according to a rule.

pub const AUTO_MODERATION_RULE_CREATE: Self = _

AutoModerationRule has been created.

pub const AUTO_MODERATION_RULE_DELETE: Self = _

AutoModerationRule has been deleted.

pub const AUTO_MODERATION_RULE_UPDATE: Self = _

AutoModerationRule has been updated.

pub const ENTITLEMENT_CREATE: Self = _

An entitlement has been created.

pub const ENTITLEMENT_DELETE: Self = _

An entitlement has been deleted.

pub const ENTITLEMENT_UPDATE: Self = _

An entitlement has been updated.

pub const BAN_ADD: Self = _

User has been banned from a guild.

pub const BAN_REMOVE: Self = _

User has been unbanned from a guild.

pub const CHANNEL_CREATE: Self = _

Channel has been created.

pub const CHANNEL_DELETE: Self = _

Channel has been deleted.

pub const CHANNEL_PINS_UPDATE: Self = _

Channel’s pins have been updated.

pub const CHANNEL_UPDATE: Self = _

Channel has been updated.

pub const COMMAND_PERMISSIONS_UPDATE: Self = _

A command’s permissions has been updated.

pub const GATEWAY_HEARTBEAT: Self = _

Heartbeat has been created.

pub const GATEWAY_HEARTBEAT_ACK: Self = _

Heartbeat has been acknowledged.

pub const GATEWAY_HELLO: Self = _

A “hello” packet has been received from the gateway.

pub const GATEWAY_INVALIDATE_SESSION: Self = _

Shard’s session has been invalidated.

A payload containing a boolean is included. If true the session is resumable. If not, then the shard must initialize a new session.

pub const GATEWAY_RECONNECT: Self = _

Gateway is indicating that a shard should perform a reconnect.

pub const GUILD_AUDIT_LOG_ENTRY_CREATE: Self = _

An audit log entry has been created.

pub const GUILD_CREATE: Self = _

A guild has been created.

pub const GUILD_DELETE: Self = _

A guild has been deleted or the current user has been removed from a guild.

pub const GUILD_EMOJIS_UPDATE: Self = _

A guild’s emojis have been updated.

pub const GUILD_INTEGRATIONS_UPDATE: Self = _

A guild’s integrations have been updated.

pub const GUILD_SCHEDULED_EVENT_CREATE: Self = _

A guild’s integrations have been updated.

pub const GUILD_SCHEDULED_EVENT_DELETE: Self = _

A guild’s integrations have been updated.

pub const GUILD_SCHEDULED_EVENT_UPDATE: Self = _

A guild’s integrations have been updated.

pub const GUILD_SCHEDULED_EVENT_USER_ADD: Self = _

A guild’s integrations have been updated.

pub const GUILD_SCHEDULED_EVENT_USER_REMOVE: Self = _

A guild’s integrations have been updated.

pub const GUILD_STICKERS_UPDATE: Self = _

A guild’s stickers have been updated.

pub const GUILD_UPDATE: Self = _

A guild has been updated.

pub const INTEGRATION_CREATE: Self = _

A guild integration was created.

pub const INTEGRATION_DELETE: Self = _

A guild integration was deleted.

pub const INTEGRATION_UPDATE: Self = _

A guild integration was updated.

pub const INTERACTION_CREATE: Self = _

An interaction was invoked by a user.

pub const INVITE_CREATE: Self = _

Invite for a channel has been created.

pub const INVITE_DELETE: Self = _

Invite for a channel has been deleted.

pub const MEMBER_ADD: Self = _

Member has been added to a guild.

pub const MEMBER_REMOVE: Self = _

Member has been removed from a guild.

pub const MEMBER_UPDATE: Self = _

Member in a guild has been updated.

pub const MEMBER_CHUNK: Self = _

Group of members from a guild.

This may be all of the remaining members or not; the chunk index in the event payload needs to be confirmed.

pub const MESSAGE_CREATE: Self = _

Message created in a channel.

pub const MESSAGE_DELETE: Self = _

Message deleted in a channel.

pub const MESSAGE_DELETE_BULK: Self = _

Multiple messages have been deleted in a channel.

pub const MESSAGE_POLL_VOTE_ADD: Self = _

Message poll vote has been added.

pub const MESSAGE_POLL_VOTE_REMOVE: Self = _

Message poll vote has been removed.

pub const MESSAGE_UPDATE: Self = _

Message in a channel has been updated.

pub const PRESENCE_UPDATE: Self = _

User’s presence details are updated.

pub const REACTION_ADD: Self = _

Reaction has been added to a message.

pub const REACTION_REMOVE: Self = _

Reaction has been removed from a message.

pub const REACTION_REMOVE_ALL: Self = _

All of the reactions for a message have been removed.

pub const REACTION_REMOVE_EMOJI: Self = _

All of a given emoji’s reactions for a message have been removed.

pub const READY: Self = _

Session is initialized.

pub const RESUMED: Self = _

Session is resumed.

pub const ROLE_CREATE: Self = _

Role has been created in a guild.

pub const ROLE_DELETE: Self = _

Role has been deleted in a guild.

pub const ROLE_UPDATE: Self = _

Role has been updated in a guild.

pub const STAGE_INSTANCE_CREATE: Self = _

Stage instance was created in a stage channel.

pub const STAGE_INSTANCE_DELETE: Self = _

Stage instance was deleted in a stage channel.

pub const STAGE_INSTANCE_UPDATE: Self = _

Stage instance was updated in a stage channel.

pub const THREAD_CREATE: Self = _

A thread has been created, relevant to the current user, or the current user has been added to a thread.

pub const THREAD_DELETE: Self = _

A thread, relevant to the current user, has been deleted.

pub const THREAD_LIST_SYNC: Self = _

The current user has gained access to a thread.

pub const THREAD_MEMBERS_UPDATE: Self = _

A user has been added to or removed from a thread.

pub const THREAD_MEMBER_UPDATE: Self = _

The thread member object for the current user has been updated.

pub const THREAD_UPDATE: Self = _

A thread has been updated.

pub const TYPING_START: Self = _

User has begun typing in a channel.

pub const UNAVAILABLE_GUILD: Self = _

Guild is unavailable, potentially due to an outage.

pub const USER_UPDATE: Self = _

Current user’s profile has been updated.

pub const VOICE_SERVER_UPDATE: Self = _

Voice server has provided an update with voice session details.

pub const VOICE_STATE_UPDATE: Self = _

User’s state in a voice channel has been updated.

pub const WEBHOOKS_UPDATE: Self = _

Webhook in a guild has been updated.

pub const AUTO_MODERATION_CONFIGURATION: Self = _

All EventTypeFlags in Intents::AUTO_MODERATION_CONFIGURATION.

pub const AUTO_MODERATION_EXECUTION: Self = _

All EventTypeFlags in Intents::AUTO_MODERATION_EXECUTION.

pub const DIRECT_MESSAGES: Self = _

All EventTypeFlags in Intents::DIRECT_MESSAGES.

pub const DIRECT_MESSAGE_REACTIONS: Self = _

All EventTypeFlags in Intents::DIRECT_MESSAGE_REACTIONS.

pub const DIRECT_MESSAGE_TYPING: Self = _

All EventTypeFlags in Intents::DIRECT_MESSAGE_TYPING.

pub const GUILDS: Self = _

All EventTypeFlags in Intents::GUILDS.

pub const GUILD_MODERATION: Self = _

All EventTypeFlags in Intents::GUILD_MODERATION.

pub const GUILD_EMOJIS_AND_STICKERS: Self = _

All EventTypeFlags in Intents::GUILD_EMOJIS_AND_STICKERS.

pub const GUILD_INTEGRATIONS: Self = _

All EventTypeFlags in Intents::GUILD_INTEGRATIONS.

pub const GUILD_INVITES: Self = _

All EventTypeFlags in Intents::GUILD_INVITES.

pub const GUILD_MEMBERS: Self = _

All EventTypeFlags in Intents::GUILD_MEMBERS.

pub const GUILD_MESSAGES: Self = _

All EventTypeFlags in Intents::GUILD_MESSAGES.

pub const MESSAGE_POLLS: Self = _

All EventTypeFlags in Intents::DIRECT_MESSAGE_POLLS and Intents::GUILD_MESSAGE_POLLS.

pub const GUILD_MESSAGE_REACTIONS: Self = _

All EventTypeFlags in Intents::GUILD_MESSAGE_REACTIONS.

pub const GUILD_MESSAGE_TYPING: Self = _

All EventTypeFlags in Intents::GUILD_MESSAGE_TYPING.

pub const GUILD_PRESENCES: Self = _

All EventTypeFlags in Intents::GUILD_PRESENCES.

pub const GUILD_SCHEDULED_EVENTS: Self = _

All EventTypeFlags in Intents::GUILD_SCHEDULED_EVENTS.

pub const GUILD_VOICE_STATES: Self = _

All EventTypeFlags in Intents::GUILD_VOICE_STATES.

pub const GUILD_WEBHOOKS: Self = _

All EventTypeFlags in Intents::GUILD_WEBHOOKS.

impl EventTypeFlags

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u128

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u128) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u128) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u128) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl EventTypeFlags

pub const fn iter(&self) -> Iter<EventTypeFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<EventTypeFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.