Module twilight_gateway::shard

Types for working with and running connections to the gateway.

At the heart of the shard module is the Shard itself: it's the interface used to start a shard, send messages to the gateway, and receive Events from it, such as new messages or channel deletions.

Once running, the shard maintains information about itself that you can obtain through it. This is information such as the latency or the current Stage of the connection, like whether it's Disconnected or Resuming the connection.

Shards are configurable through the ShardConfigBuilder struct, which provides a clean interface for correctly building a ShardConfig.

Re-exports

pub use self::config::ShardConfig;

pub use self::stage::Stage;

Modules

config

stage
Utilities for knowing and parsing the current connection stage of a shard.

Structs

Events
A stream of events from a Shard.

Information
Information about a shard, including its latency, current session sequence, and connection stage.

Latency
Information about the latency of a Shard's websocket connection.

ResumeSession
Holds the sessions id and sequence number to resume this shard's session with with

SessionInactiveError
The Shard's session is inactive.

Shard

ShardSink
A sink which tungstenite messages can be sunk into. ⚓

Enums

CommandError
Sending a command failed.

ShardStartError
Starting a shard and connecting to the gateway failed.