Struct twilight_gateway::shard::ShardBuilder

pub struct ShardBuilder(_);

Builder to configure and construct a shard.

Use ShardBuilder::new to start configuring a new Shard.

Examples

Create a new shard, setting the large_threshold to 100 and the shard ID to 5 out of 10:

use std::env;
use twilight_gateway::{Intents, Shard};

let token = env::var("DISCORD_TOKEN")?;

let shard = Shard::builder(token, Intents::GUILD_MESSAGE_REACTIONS)
    .large_threshold(100)?
    .shard(5, 10)?
    .build();

Implementations

impl ShardBuilder

pub fn new(token: impl Into<String>, intents: Intents) -> Self

Create a new builder to configure and construct a shard.

Refer to each method to learn their default values.

pub fn build(self) -> (Shard, Events)

Consume the builder, constructing a shard.

pub const fn event_types(self, event_types: EventTypeFlags) -> Self

Set the event types to process.

This is an optimization technique; all events not included in the provided event type flags will not be deserialized by the gateway and will be discarded. All events will still be sent if EventTypeFlags::SHARD_PAYLOAD is enabled.

pub fn gateway_url(self, gateway_url: Option<String>) -> Self

Set the URL used for connecting to Discord’s gateway

pub fn http_client(self, http_client: HttpClient) -> Self

Set the HTTP client to be used by the shard for getting gateway information.

Default is a new, unconfigured instance of an HTTP client.

pub fn large_threshold(
self,
large_threshold: u64
) -> Result<Self, LargeThresholdError>

Set the maximum number of members in a guild to load the member list.

Default value is 250. The minimum value is 50 and the maximum is 250.

Examples

If you pass 200, then if there are 250 members in a guild the member list won’t be sent. If there are 150 members, then the list will be sent.

Errors

Returns a LargeThresholdErrorType::TooFew error type if the provided value is below 50.

Returns a LargeThresholdErrorType::TooMany error type if the provided value is above 250.

pub fn presence(self, presence: UpdatePresencePayload) -> Self

Set the presence to use automatically when starting a new session.

Default is no presence, which defaults to strictly being “online” with no special qualities.

Examples

Set the bot user’s presence to idle with the status “Not accepting commands”:

use twilight_gateway::{Intents, Shard};
use twilight_model::gateway::{
    payload::update_presence::UpdatePresencePayload,
    presence::{ActivityType, MinimalActivity, Status},
};

let shard = Shard::builder("token", Intents::empty())
    .presence(UpdatePresencePayload::new(
        vec![MinimalActivity {
            kind: ActivityType::Playing,
            name: "Not accepting commands".into(),
            url: None,
        }
        .into()],
        false,
        None,
        Status::Idle,
    )?);

pub fn queue(self, queue: Arc<dyn Queue>) -> Self

Set the queue to use for queueing shard connections.

You probably don’t need to set this yourself, because the Cluster manages that for you. Refer to the queue module for more information.

The default value is a queue used only by this shard, or a queue used by all shards when ran by a Cluster.

pub fn shard(
self,
shard_id: u64,
shard_total: u64
) -> Result<Self, ShardIdError>

Set the shard ID to connect as, and the total number of shards used by the bot.

The shard ID is 0-indexed, while the total is 1-indexed.

The default value is a shard ID of 0 and a shard total of 1, which is good for smaller bots.

Note: If your bot is in over 250’000 guilds then shard_total should probably be a multiple of 16 if you’re in the “Large Bot Sharding” program.

Examples

If you have 19 shards, then your last shard will have an ID of 18 out of a total of 19 shards:

use twilight_gateway::{Intents, Shard};
use std::env;

let token = env::var("DISCORD_TOKEN")?;

let shard = Shard::builder(token, Intents::empty()).shard(18, 19)?.build();

Errors

Returns a ShardIdErrorType::IdTooLarge error type if the shard ID to connect as is larger than the total.