Trait twilight_gateway::shard::Command

pub trait Command: Sealed { }

Trait marker to denote what can be provided to Shard::command.

This trait exists to make it easier to determine what commands can be sent to the Discord Gateway API.

To send an arbitrary command to the Discord Gateway API then Shard::send may be used.