Struct twilight_gateway::EventTypeFlags

pub struct EventTypeFlags { /* private fields */ }

Bitflags representing all of the possible types of events.

Implementations

impl EventTypeFlags

pub const BAN_ADD: Self = Self{bits: 1,}

User has been banned from a guild.

pub const BAN_REMOVE: Self = Self{bits: 1 << 1,}

User has been unbanned from a guild.

pub const CHANNEL_CREATE: Self = Self{bits: 1 << 2,}

Channel has been created.

pub const CHANNEL_DELETE: Self = Self{bits: 1 << 3,}

Channel has been deleted.

pub const CHANNEL_PINS_UPDATE: Self = Self{bits: 1 << 4,}

Channel’s pins have been updated.

pub const CHANNEL_UPDATE: Self = Self{bits: 1 << 5,}

Channel has been updated.

pub const GATEWAY_HEARTBEAT: Self = Self{bits: 1 << 6,}

Heartbeat has been created.

pub const GATEWAY_HEARTBEAT_ACK: Self = Self{bits: 1 << 7,}

Heartbeat has been acknowledged.

pub const GATEWAY_HELLO: Self = Self{bits: 1 << 8,}

A “hello” packet has been received from the gateway.

pub const GATEWAY_INVALIDATE_SESSION: Self = Self{bits: 1 << 8,}

Shard’s session has been invalidated.

A payload containing a boolean is included. If true the session is resumable. If not, then the shard must initialize a new session.

pub const GATEWAY_RECONNECT: Self = Self{bits: 1 << 9,}

Gateway is indicating that a shard should perform a reconnect.

pub const GIFT_CODE_UPDATE: Self = Self{bits: 1 << 49,}

Gift code sent in a channel has been updated.

pub const GUILD_CREATE: Self = Self{bits: 1 << 10,}

A guild has been created.

pub const GUILD_DELETE: Self = Self{bits: 1 << 11,}

A guild has been deleted or the current user has been removed from a guild.

pub const GUILD_EMOJIS_UPDATE: Self = Self{bits: 1 << 12,}

A guild’s emojis have been updated.

pub const GUILD_INTEGRATIONS_UPDATE: Self = Self{bits: 1 << 13,}

A guild’s integrations have been updated.

pub const GUILD_STICKERS_UPDATE: Self = Self{bits: 1 << 63,}

A guild’s stickers have been updated.

pub const GUILD_UPDATE: Self = Self{bits: 1 << 14,}

A guild has been updated.

pub const INTEGRATION_CREATE: Self = Self{bits: 1 << 60,}

A guild integration was created.

pub const INTEGRATION_DELETE: Self = Self{bits: 1 << 61,}

A guild integration was deleted.

pub const INTEGRATION_UPDATE: Self = Self{bits: 1 << 62,}

A guild integration was updated.

pub const INTERACTION_CREATE: Self = Self{bits: 1 << 56,}

An interaction was invoked by a user.

pub const INVITE_CREATE: Self = Self{bits: 1 << 46,}

Invite for a channel has been created.

pub const INVITE_DELETE: Self = Self{bits: 1 << 47,}

Invite for a channel has been deleted.

pub const MEMBER_ADD: Self = Self{bits: 1 << 15,}

Member has been added to a guild.

pub const MEMBER_REMOVE: Self = Self{bits: 1 << 16,}

Member has been removed from a guild.

pub const MEMBER_UPDATE: Self = Self{bits: 1 << 17,}

Member in a guild has been updated.

pub const MEMBER_CHUNK: Self = Self{bits: 1 << 18,}

Group of members from a guild.

This may be all of the remaining members or not; the chunk index in the event payload needs to be confirmed.

pub const MESSAGE_CREATE: Self = Self{bits: 1 << 19,}

Message created in a channel.

pub const MESSAGE_DELETE: Self = Self{bits: 1 << 20,}

Message deleted in a channel.

pub const MESSAGE_DELETE_BULK: Self = Self{bits: 1 << 21,}

Multiple messages have been deleted in a channel.

pub const MESSAGE_UPDATE: Self = Self{bits: 1 << 22,}

Message in a channel has been updated.

pub const PRESENCE_UPDATE: Self = Self{bits: 1 << 23,}

User’s presence details are updated.

pub const PRESENCES_REPLACE: Self = Self{bits: 1 << 24,}

Group of presences are replaced.

This is a placeholder as it can happen for bots but has no real meaning.

pub const REACTION_ADD: Self = Self{bits: 1 << 25,}

Reaction has been added to a message.

pub const REACTION_REMOVE: Self = Self{bits: 1 << 26,}

Reaction has been removed from a message.

pub const REACTION_REMOVE_ALL: Self = Self{bits: 1 << 27,}

All of the reactions for a message have been removed.

pub const REACTION_REMOVE_EMOJI: Self = Self{bits: 1 << 48,}

All of a given emoji’s reactions for a message have been removed.

pub const READY: Self = Self{bits: 1 << 28,}

Session is initialized.

pub const RESUMED: Self = Self{bits: 1 << 29,}

Session is resumed.

pub const ROLE_CREATE: Self = Self{bits: 1 << 30,}

Role has been created in a guild.

pub const ROLE_DELETE: Self = Self{bits: 1 << 31,}

Role has been deleted in a guild.

pub const ROLE_UPDATE: Self = Self{bits: 1 << 32,}

Role has been updated in a guild.

pub const SHARD_CONNECTED: Self = Self{bits: 1 << 33,}

Shard has finalized a session with the gateway.

pub const SHARD_CONNECTING: Self = Self{bits: 1 << 34,}

Shard has begun connecting to the gateway.

pub const SHARD_DISCONNECTED: Self = Self{bits: 1 << 35,}

Shard has disconnected from the gateway.

pub const SHARD_IDENTIFYING: Self = Self{bits: 1 << 36,}

Shard is identifying to create a session with the gateway.

pub const SHARD_PAYLOAD: Self = Self{bits: 1 << 45,}

Incoming message has been received from the gateway.

pub const SHARD_RECONNECTING: Self = Self{bits: 1 << 37,}

Shard is reconnecting to the gateway.

pub const SHARD_RESUMING: Self = Self{bits: 1 << 38,}

Shard is resuming a session with the gateway.

pub const STAGE_INSTANCE_CREATE: Self = Self{bits: 1 << 57,}

Stage instance was created in a stage channel.

pub const STAGE_INSTANCE_DELETE: Self = Self{bits: 1 << 58,}

Stage instance was deleted in a stage channel.

pub const STAGE_INSTANCE_UPDATE: Self = Self{bits: 1 << 59,}

Stage instance was updated in a stage channel.

pub const THREAD_CREATE: Self = Self{bits: 1 << 50,}

A thread has been created, relevant to the current user, or the current user has been added to a thread.

pub const THREAD_DELETE: Self = Self{bits: 1 << 52,}

A thread, relevant to the current user, has been deleted.

pub const THREAD_LIST_SYNC: Self = Self{bits: 1 << 53,}

The current user has gained access to a thread.

pub const THREAD_MEMBERS_UPDATE: Self = Self{bits: 1 << 55,}

A user has been added to or removed from a thread.

pub const THREAD_MEMBER_UPDATE: Self = Self{bits: 1 << 54,}

The thread member object for the current user has been updated.

pub const THREAD_UPDATE: Self = Self{bits: 1 << 51,}

A thread has been updated.

pub const TYPING_START: Self = Self{bits: 1 << 39,}

User has begun typing in a channel.

pub const UNAVAILABLE_GUILD: Self = Self{bits: 1 << 40,}

Guild is unavailable, potentially due to an outage.

pub const USER_UPDATE: Self = Self{bits: 1 << 41,}

Current user’s profile has been updated.

pub const VOICE_SERVER_UPDATE: Self = Self{bits: 1 << 42,}

Voice server has provided an update with voice session details.

pub const VOICE_STATE_UPDATE: Self = Self{bits: 1 << 43,}

User’s state in a voice channel has been updated.

pub const WEBHOOKS_UPDATE: Self = Self{bits: 1 << 44,}

Webhook in a guild has been updated.

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl EventTypeFlags

pub const DIRECT_MESSAGES: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::MESSAGE_CREATE.bits() | EventTypeFlags::MESSAGE_DELETE.bits() | EventTypeFlags::MESSAGE_DELETE_BULK.bits() | EventTypeFlags::MESSAGE_UPDATE.bits(), )

All EventTypeFlags in Intents::DIRECT_MESSAGES.

pub const DIRECT_MESSAGE_REACTIONS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::REACTION_ADD.bits() | EventTypeFlags::REACTION_REMOVE.bits() | EventTypeFlags::REACTION_REMOVE_ALL.bits() | EventTypeFlags::REACTION_REMOVE_EMOJI.bits(), )

All EventTypeFlags in Intents::DIRECT_MESSAGE_REACTIONS.

pub const DIRECT_MESSAGE_TYPING: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::TYPING_START.bits())

All EventTypeFlags in Intents::DIRECT_MESSAGE_TYPING.

pub const GUILDS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::CHANNEL_CREATE.bits() | EventTypeFlags::CHANNEL_DELETE.bits() | EventTypeFlags::CHANNEL_PINS_UPDATE.bits() | EventTypeFlags::CHANNEL_UPDATE.bits() | EventTypeFlags::GUILD_CREATE.bits() | EventTypeFlags::GUILD_DELETE.bits() | EventTypeFlags::GUILD_UPDATE.bits() | EventTypeFlags::ROLE_CREATE.bits() | EventTypeFlags::ROLE_DELETE.bits() | EventTypeFlags::ROLE_UPDATE.bits(), )

All EventTypeFlags in Intents::GUILDS.

pub const GUILD_BANS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::BAN_ADD.bits() | EventTypeFlags::BAN_REMOVE.bits(), )

All EventTypeFlags in Intents::GUILD_BANS.

pub const GUILD_EMOJIS_AND_STICKERS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::GUILD_EMOJIS_UPDATE.bits() | EventTypeFlags::GUILD_STICKERS_UPDATE.bits(), )

All EventTypeFlags in Intents::GUILD_EMOJIS_AND_STICKERS.

pub const GUILD_INTEGRATIONS: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::GUILD_INTEGRATIONS_UPDATE.bits())

All EventTypeFlags in Intents::GUILD_INTEGRATIONS.

pub const GUILD_INVITES: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::INVITE_CREATE.bits() | EventTypeFlags::INVITE_DELETE.bits(), )

All EventTypeFlags in Intents::GUILD_INVITES.

pub const GUILD_MEMBERS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::MEMBER_ADD.bits() | EventTypeFlags::MEMBER_REMOVE.bits() | EventTypeFlags::MEMBER_UPDATE.bits(), )

All EventTypeFlags in Intents::GUILD_MEMBERS.

pub const GUILD_MESSAGES: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::MESSAGE_CREATE.bits() | EventTypeFlags::MESSAGE_DELETE.bits() | EventTypeFlags::MESSAGE_DELETE.bits() | EventTypeFlags::MESSAGE_DELETE_BULK.bits(), )

All EventTypeFlags in Intents::GUILD_MESSAGES.

pub const GUILD_MESSAGE_REACTIONS: EventTypeFlags = EventTypeFlags::from_bits_truncate( EventTypeFlags::REACTION_ADD.bits() | EventTypeFlags::REACTION_REMOVE.bits() | EventTypeFlags::REACTION_REMOVE_ALL.bits() | EventTypeFlags::REACTION_REMOVE_EMOJI.bits(), )

All EventTypeFlags in Intents::GUILD_MESSAGE_REACTIONS.

pub const GUILD_MESSAGE_TYPING: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::TYPING_START.bits())

All EventTypeFlags in Intents::GUILD_MESSAGE_TYPING.

pub const GUILD_PRESENCES: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::PRESENCE_UPDATE.bits())

All EventTypeFlags in Intents::GUILD_PRESENCES.

pub const GUILD_VOICE_STATES: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::VOICE_STATE_UPDATE.bits())

All EventTypeFlags in Intents::GUILD_VOICE_STATES.

pub const GUILD_WEBHOOKS: EventTypeFlags = EventTypeFlags::from_bits_truncate(EventTypeFlags::WEBHOOKS_UPDATE.bits())

All EventTypeFlags in Intents::GUILD_WEBHOOKS.