Struct twilight_gateway::EventTypeFlags

#[non_exhaustive]
pub struct EventTypeFlags { /* fields omitted */ }

Bitflags representing all of the possible types of events.

Implementations

impl EventTypeFlags

pub const BAN_ADD: EventTypeFlags

User has been banned from a guild.

pub const BAN_REMOVE: EventTypeFlags

User has been unbanned from a guild.

pub const CHANNEL_CREATE: EventTypeFlags

Channel has been created.

pub const CHANNEL_DELETE: EventTypeFlags

Channel has been deleted.

pub const CHANNEL_PINS_UPDATE: EventTypeFlags

Channel’s pins have been updated.

pub const CHANNEL_UPDATE: EventTypeFlags

Channel has been updated.

pub const GATEWAY_HEARTBEAT: EventTypeFlags

Heartbeat has been created.

pub const GATEWAY_HEARTBEAT_ACK: EventTypeFlags

Heartbeat has been acknowledged.

pub const GATEWAY_HELLO: EventTypeFlags

A “hello” packet has been received from the gateway.

pub const GATEWAY_INVALIDATE_SESSION: EventTypeFlags

Shard’s session has been invalidated.

A payload containing a boolean is included. If true the session is resumable. If not, then the shard must initialize a new session.

pub const GATEWAY_RECONNECT: EventTypeFlags

Gateway is indicating that a shard should perform a reconnect.

pub const GIFT_CODE_UPDATE: EventTypeFlags

Gift code sent in a channel has been updated.

pub const GUILD_CREATE: EventTypeFlags

A guild has been created.

pub const GUILD_DELETE: EventTypeFlags

A guild has been deleted or the current user has been removed from a guild.

pub const GUILD_EMOJIS_UPDATE: EventTypeFlags

A guild’s emojis have been updated.

pub const GUILD_INTEGRATIONS_UPDATE: EventTypeFlags

A guild’s integrations have been updated.

pub const GUILD_UPDATE: EventTypeFlags

A guild has been updated.

pub const INVITE_CREATE: EventTypeFlags

Invite for a channel has been created.

pub const INVITE_DELETE: EventTypeFlags

Invite for a channel has been deleted.

pub const MEMBER_ADD: EventTypeFlags

Member has been added to a guild.

pub const MEMBER_REMOVE: EventTypeFlags

Member has been removed from a guild.

pub const MEMBER_UPDATE: EventTypeFlags

Member in a guild has been updated.

pub const MEMBER_CHUNK: EventTypeFlags

Group of members from a guild.

This may be all of the remaining members or not; the chunk index in the event payload needs to be confirmed.

pub const MESSAGE_CREATE: EventTypeFlags

Message created in a channel.

pub const MESSAGE_DELETE: EventTypeFlags

Message deleted in a channel.

pub const MESSAGE_DELETE_BULK: EventTypeFlags

Multiple messages have been deleted in a channel.

pub const MESSAGE_UPDATE: EventTypeFlags

Message in a channel has been updated.

pub const PRESENCE_UPDATE: EventTypeFlags

User’s presence details are updated.

pub const PRESENCES_REPLACE: EventTypeFlags

Group of presences are replaced.

This is a placeholder as it can happen for bots but has no real meaning.

pub const REACTION_ADD: EventTypeFlags

Reaction has been added to a message.

pub const REACTION_REMOVE: EventTypeFlags

Reaction has been removed from a message.

pub const REACTION_REMOVE_ALL: EventTypeFlags

All of the reactions for a message have been removed.

pub const REACTION_REMOVE_EMOJI: EventTypeFlags

All of a given emoji’s reactions for a message have been removed.

pub const READY: EventTypeFlags

Session is initialized.

pub const RESUMED: EventTypeFlags

Session is resumed.

pub const ROLE_CREATE: EventTypeFlags

Role has been created in a guild.

pub const ROLE_DELETE: EventTypeFlags

Role has been deleted in a guild.

pub const ROLE_UPDATE: EventTypeFlags

Role has been updated in a guild.

pub const SHARD_CONNECTED: EventTypeFlags

Shard has finalized a session with the gateway.

pub const SHARD_CONNECTING: EventTypeFlags

Shard has begun connecting to the gateway.

pub const SHARD_DISCONNECTED: EventTypeFlags

Shard has disconnected from the gateway.

pub const SHARD_IDENTIFYING: EventTypeFlags

Shard is identifying to create a session with the gateway.

pub const SHARD_PAYLOAD: EventTypeFlags

Incoming message has been received from the gateway.

pub const SHARD_RECONNECTING: EventTypeFlags

Shard is reconnecting to the gateway.

pub const SHARD_RESUMING: EventTypeFlags

Shard is resuming a session with the gateway.

pub const TYPING_START: EventTypeFlags

User has begun typing in a channel.

pub const UNAVAILABLE_GUILD: EventTypeFlags

Guild is unavailable, potentially due to an outage.

pub const USER_UPDATE: EventTypeFlags

Current user’s profile has been updated.

pub const VOICE_SERVER_UPDATE: EventTypeFlags

Voice server has provided an update with voice session details.

pub const VOICE_STATE_UPDATE: EventTypeFlags

User’s state in a voice channel has been updated.

pub const WEBHOOKS_UPDATE: EventTypeFlags

Webhook in a guild has been updated.

pub const fn empty() -> EventTypeFlags

Returns an empty set of flags

pub const fn all() -> EventTypeFlags

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<EventTypeFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> EventTypeFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> EventTypeFlags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: EventTypeFlags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: EventTypeFlags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: EventTypeFlags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: EventTypeFlags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: EventTypeFlags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: EventTypeFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.