Struct twilight_gateway::EventTypeFlags

pub struct EventTypeFlags { /* fields omitted */ }

Bitflags representing all of the possible types of events.

Implementations

impl EventTypeFlags

pub const BAN_ADD: EventTypeFlags

A user was banned from a guild.

pub const BAN_REMOVE: EventTypeFlags

A user was unbanned from a guild.

pub const CHANNEL_CREATE: EventTypeFlags

A channel was created.

pub const CHANNEL_DELETE: EventTypeFlags

A channel was deleted.

pub const CHANNEL_PINS_UPDATE: EventTypeFlags

A channel's pins were updated.

pub const CHANNEL_UPDATE: EventTypeFlags

A channel was updated.

pub const GATEWAY_HEARTBEAT: EventTypeFlags

A heartbeat was created.

pub const GATEWAY_HEARTBEAT_ACK: EventTypeFlags

A heartbeat was acknowledged.

pub const GATEWAY_HELLO: EventTypeFlags

A "hello" packet was received from the gateway.

pub const GATEWAY_INVALIDATE_SESSION: EventTypeFlags

A shard's session was invalidated.

true if resumeable. If not, then the shard must do a full reconnect.

pub const GATEWAY_RECONNECT: EventTypeFlags

The gateway is indicating to perform a reconnect.

pub const GIFT_CODE_UPDATE: EventTypeFlags

pub const GUILD_CREATE: EventTypeFlags

A guild was created.

pub const GUILD_DELETE: EventTypeFlags

A guild was deleted or the current user was removed from a guild.

pub const GUILD_EMOJIS_UPDATE: EventTypeFlags

A guild's emojis were updated.

pub const GUILD_INTEGRATIONS_UPDATE: EventTypeFlags

A guild's integrations were updated.

pub const GUILD_UPDATE: EventTypeFlags

A guild was updated.

pub const INVITE_CREATE: EventTypeFlags

pub const INVITE_DELETE: EventTypeFlags

pub const MEMBER_ADD: EventTypeFlags

pub const MEMBER_REMOVE: EventTypeFlags

pub const MEMBER_UPDATE: EventTypeFlags

pub const MEMBER_CHUNK: EventTypeFlags

pub const MESSAGE_CREATE: EventTypeFlags

pub const MESSAGE_DELETE: EventTypeFlags

pub const MESSAGE_DELETE_BULK: EventTypeFlags

pub const MESSAGE_UPDATE: EventTypeFlags

pub const PRESENCE_UPDATE: EventTypeFlags

pub const PRESENCES_REPLACE: EventTypeFlags

pub const REACTION_ADD: EventTypeFlags

pub const REACTION_REMOVE: EventTypeFlags

pub const REACTION_REMOVE_ALL: EventTypeFlags

pub const REACTION_REMOVE_EMOJI: EventTypeFlags

pub const READY: EventTypeFlags

pub const RESUMED: EventTypeFlags

pub const ROLE_CREATE: EventTypeFlags

pub const ROLE_DELETE: EventTypeFlags

pub const ROLE_UPDATE: EventTypeFlags

pub const SHARD_CONNECTED: EventTypeFlags

pub const SHARD_CONNECTING: EventTypeFlags

pub const SHARD_DISCONNECTED: EventTypeFlags

pub const SHARD_IDENTIFYING: EventTypeFlags

pub const SHARD_PAYLOAD: EventTypeFlags

pub const SHARD_RECONNECTING: EventTypeFlags

pub const SHARD_RESUMING: EventTypeFlags

pub const TYPING_START: EventTypeFlags

pub const UNAVAILABLE_GUILD: EventTypeFlags

pub const USER_UPDATE: EventTypeFlags

pub const VOICE_SERVER_UPDATE: EventTypeFlags

pub const VOICE_STATE_UPDATE: EventTypeFlags

pub const WEBHOOKS_UPDATE: EventTypeFlags

pub const fn empty() -> EventTypeFlags

Returns an empty set of flags

pub const fn all() -> EventTypeFlags

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<EventTypeFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> EventTypeFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> EventTypeFlags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: EventTypeFlags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: EventTypeFlags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: EventTypeFlags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: EventTypeFlags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: EventTypeFlags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: EventTypeFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.